+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project attempts to adhere to [Semantic Versioning](http://semver.org/)
 since `1.7.0`.
 
 ## [Unreleased]
+
+## [1.13.0] - 2019-07-05
 ### Added
 - You can now set the `TCOD_RENDERER` and `TCOD_VSYNC` environment variables to
   force specific options to be used.